# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Taiwan Tourism Open Data API
  description: '

    以 **交通部觀光署「景點 - 觀光資訊資料庫」** 政府開放資料（6,000+ 筆全台景點）為基礎的 RESTful API。


    資料來源：<https://data.gov.tw/dataset/7777>（觀光資料標準 V2.1）


    ### 資源 (Resources)

    | Resource | 說明 |

    |---|---|

    | `attractions` | 景點（支援關鍵字搜尋、多條件篩選、排序、分頁、附近景點查詢） |

    | `attractions/{id}/images` | 景點圖片（子資源） |

    | `categories` | 景點類型（文化類、生態類…） |

    | `cities` / `towns` | 縣市 / 鄉鎮市區 |

    | `reviews` | 遊客評論（1~5 星） |

    | `trips` / `trips/{id}/items` | 旅遊行程與行程項目 |

    | `stats` | 統計資訊（唯讀） |


    ### 驗證

    所有 **讀取 (GET)** 操作公開；**新增 / 修改 / 刪除 (POST / PUT / PATCH / DELETE)** 需在 Header
    帶入

    `X-API-Key: ntub-iot-2026`（可按右上角 **Authorize** 按鈕設定）。


    ### 慣例

    * 列表回應格式：`{"total", "page", "page_size", "pages", "items"}`，並附 `X-Total-Count`
    header

    * 建立成功回傳 `201 Created` 與 `Location` header；刪除成功回傳 `204 No Content`

    * 錯誤格式：`{"detail": "錯誤說明"}`（404 查無資料、409 資料衝突、422 驗證失敗）

    '
  contact:
    name: NTUB IoT App 作業
  license:
    name: 政府資料開放授權條款－第1版
    url: https://data.gov.tw/license
  version: 1.0.0
paths:
  /api/v1/attractions:
    get:
      tags:
      - Attractions 景點
      summary: 查詢景點列表（分頁、篩選、排序）
      operationId: list_attractions_api_v1_attractions_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 關鍵字（比對名稱、別名、描述、地址）
          examples:
          - 老街
          title: Q
        description: 關鍵字（比對名稱、別名、描述、地址）
      - name: city_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 縣市代碼
          examples:
          - '63000'
          title: City Code
        description: 縣市代碼
      - name: city
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 縣市名稱（部分比對）
          examples:
          - 臺北
          title: City
        description: 縣市名稱（部分比對）
      - name: town_code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 鄉鎮市區代碼
          title: Town Code
        description: 鄉鎮市區代碼
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: integer
          - type: 'null'
          description: 景點類型代碼，可重複帶入（OR 條件）
          title: Category Id
        description: 景點類型代碼，可重複帶入（OR 條件）
      - name: is_free
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: 是否免費入場
          title: Is Free
        description: 是否免費入場
      - name: service_status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ServiceStatus'
          - type: 'null'
          description: 營運狀態
          title: Service Status
        description: 營運狀態
      - name: has_image
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: 是否有圖片
          title: Has Image
        description: 是否有圖片
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 標籤（完全比對）
          examples:
          - 賞楓
          title: Tag
        description: 標籤（完全比對）
      - name: sort
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/SortField'
          description: 排序欄位，前面加 - 代表遞減
          default: id
        description: 排序欄位，前面加 - 代表遞減
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 頁碼（從 1 開始）
          default: 1
          title: Page
        description: 頁碼（從 1 開始）
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 每頁筆數（1~100）
          default: 20
          title: Page Size
        description: 每頁筆數（1~100）
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AttractionSummary_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Attractions 景點
      summary: 新增景點
      operationId: create_attraction_api_v1_attractions_post
      security:
      - APIKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttractionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttractionDetail'
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Conflict
        '422':
          description: 欄位驗證失敗或參照的代碼不存在
  /api/v1/attractions/nearby:
    get:
      tags:
      - Attractions 景點
      summary: 查詢附近景點（依距離排序）
      operationId: nearby_attractions_api_v1_attractions_nearby_get
      parameters:
      - name: lat
        in: query
        required: true
        schema:
          type: number
          maximum: 90
          minimum: -90
          description: 緯度
          examples:
          - 25.0421
          title: Lat
        description: 緯度
      - name: lon
        in: query
        required: true
        schema:
          type: number
          maximum: 180
          minimum: -180
          description: 經度
          examples:
          - 121.5253
          title: Lon
        description: 經度
      - name: radius_km
        in: query
        required: false
        schema:
          type: number
          maximum: 50
          exclusiveMinimum: 0
          description: 搜尋半徑（公里）
          default: 3
          title: Radius Km
        description: 搜尋半徑（公里）
      - name: category_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 景點類型代碼
          title: Category Id
        description: 景點類型代碼
      - name: is_free
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Free
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttractionNearby'
                title: Response Nearby Attractions Api V1 Attractions Nearby Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attractions/{attraction_id}:
    get:
      tags:
      - Attractions 景點
      summary: 取得單一景點詳細資料
      operationId: get_attraction_api_v1_attractions__attraction_id__get
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttractionDetail'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Attractions 景點
      summary: 整筆更新景點（未提供的欄位會被重設為預設值）
      operationId: replace_attraction_api_v1_attractions__attraction_id__put
      security:
      - APIKeyHeader: []
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttractionFields'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttractionDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Attractions 景點
      summary: 部分更新景點（只更新有傳入的欄位）
      operationId: update_attraction_api_v1_attractions__attraction_id__patch
      security:
      - APIKeyHeader: []
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttractionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttractionDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Attractions 景點
      summary: 刪除景點（連同其圖片、評論、行程項目）
      operationId: delete_attraction_api_v1_attractions__attraction_id__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attractions/{attraction_id}/images:
    get:
      tags:
      - Attraction Images 景點圖片
      summary: 列出景點的所有圖片
      operationId: list_images_api_v1_attractions__attraction_id__images_get
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ImageOut'
                title: Response List Images Api V1 Attractions  Attraction Id  Images
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Attraction Images 景點圖片
      summary: 新增景點圖片
      operationId: create_image_api_v1_attractions__attraction_id__images_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/attractions/{attraction_id}/images/{image_id}:
    get:
      tags:
      - Attraction Images 景點圖片
      summary: 取得單張圖片
      operationId: get_image_api_v1_attractions__attraction_id__images__image_id__get
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Attraction Images 景點圖片
      summary: 整筆更新圖片
      operationId: replace_image_api_v1_attractions__attraction_id__images__image_id__put
      security:
      - APIKeyHeader: []
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Attraction Images 景點圖片
      summary: 部分更新圖片
      operationId: update_image_api_v1_attractions__attraction_id__images__image_id__patch
      security:
      - APIKeyHeader: []
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Attraction Images 景點圖片
      summary: 刪除圖片
      operationId: delete_image_api_v1_attractions__attraction_id__images__image_id__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: attraction_id
        in: path
        required: true
        schema:
          type: string
          title: Attraction Id
      - name: image_id
        in: path
        required: true
        schema:
          type: integer
          title: Image Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AttractionCreate:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 1
          title: Name
          examples:
          - 北科大校園
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - 位於臺北市中正區的科技大學
        alternate_names:
          items:
            type: string
          type: array
          title: Alternate Names
          examples:
          - - 臺北商大
        latitude:
          anyOf:
          - type: number
            maximum: 90.0
            minimum: -90.0
          - type: 'null'
          title: Latitude
          examples:
          - 25.0421
        longitude:
          anyOf:
          - type: number
            maximum: 180.0
            minimum: -180.0
          - type: 'null'
          title: Longitude
          examples:
          - 121.5253
        city_code:
          anyOf:
          - type: string
          - type: 'null'
          title: City Code
          examples:
          - '63000'
        town_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Town Code
          examples:
          - '63000050'
        zip_code:
          anyOf:
          - type: string
            maxLength: 10
          - type: 'null'
          title: Zip Code
          examples:
          - '100'
        street_address:
          anyOf:
          - type: string
            maxLength: 300
          - type: 'null'
          title: Street Address
          examples:
          - 濟南路一段321號
        telephones:
          items:
            type: string
          type: array
          title: Telephones
          examples:
          - - (02)2322-6050
        website_url:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Website Url
          examples:
          - https://www.ntub.edu.tw
        service_time_info:
          anyOf:
          - type: string
          - type: 'null'
          title: Service Time Info
          examples:
          - 週一至週五 08:00-22:00
        traffic_info:
          anyOf:
          - type: string
          - type: 'null'
          title: Traffic Info
        parking_info:
          anyOf:
          - type: string
          - type: 'null'
          title: Parking Info
        fee_info:
          anyOf:
          - type: string
          - type: 'null'
          title: Fee Info
        service_status:
          $ref: '#/components/schemas/ServiceStatus'
          description: 0 永久停止 / 1 正常營運 / 2 非營運時段 / 3 暫時停止營運 / 9 待確認
          default: 1
        is_public_access:
          type: boolean
          title: Is Public Access
          default: true
        is_free:
          type: boolean
          title: Is Free
          default: false
          examples:
          - true
        visit_duration:
          anyOf:
          - type: integer
            maximum: 100000.0
            minimum: 0.0
          - type: 'null'
          title: Visit Duration
          description: 建議停留時間（分鐘）
          examples:
          - 60
        assets_class:
          anyOf:
          - type: integer
          - type: 'null'
          title: Assets Class
          description: 文化資產級別代碼 (0-5, 9)
        tags:
          items:
            type: string
          type: array
          title: Tags
          examples:
          - - 校園
            - 學習
        remarks:
          anyOf:
          - type: string
          - type: 'null'
          title: Remarks
        category_ids:
          items:
            type: integer
          type: array
          title: Category Ids
          description: 景點類型代碼列表
          examples:
          - - 1
            - 12
        id:
          anyOf:
          - type: string
            maxLength: 40
            minLength: 1
            pattern: ^[A-Za-z0-9_\-]+$
          - type: 'null'
          title: Id
          description: 景點 ID；省略時自動產生（格式 Attraction_USER_xxxxxxxx）
      type: object
      required:
      - name
      title: AttractionCreate
...
    AttractionNearby:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        city:
          anyOf:
          - $ref: '#/components/schemas/CityRef'
          - type: 'null'
        town:
          anyOf:
          - $ref: '#/components/schemas/TownRef'
          - type: 'null'
        categories:
          items:
            $ref: '#/components/schemas/CategoryRef'
          type: array
          title: Categories
          default: []
        latitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Latitude
        longitude:
          anyOf:
          - type: number
          - type: 'null'
          title: Longitude
        is_free:
          type: boolean
          title: Is Free
        service_status:
          type: integer
          title: Service Status
        cover_image:
          anyOf:
          - type: string
          - type: 'null'
          title: Cover Image
        distance_km:
          type: number
          title: Distance Km
      type: object
      required:
      - id
      - name
      - is_free
      - service_status
      - distance_km
      title: AttractionNearby
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Message:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: Message
...
    Page_AttractionSummary_:
      properties:
        total:
          type: integer
          title: Total
          description: 符合條件的總筆數
        page:
          type: integer
          title: Page
          description: 目前頁碼（從 1 開始）
        page_size:
          type: integer
          title: Page Size
          description: 每頁筆數
        pages:
          type: integer
          title: Pages
          description: 總頁數
        items:
          items:
            $ref: '#/components/schemas/AttractionSummary'
          type: array
          title: Items
      type: object
      required:
      - total
      - page
      - page_size
      - pages
      - items
      title: Page[AttractionSummary]
...
    ServiceStatus:
      type: integer
      enum:
      - 0
      - 1
      - 2
      - 3
      - 9
      title: ServiceStatus
      description: 9. 營運服務狀態代碼 (ServiceStatusEnum)
    SortField:
      type: string
      enum:
      - id
      - name
      - -name
      - updated_at
      - -updated_at
      - source_update_time
      - -source_update_time
      title: SortField
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      description: 寫入操作需帶入 API Key（預設為 `ntub-iot-2026`，可用環境變數 `API_KEY` 修改）
      in: header
      name: X-API-Key
...
servers:
- url: http://127.0.0.1:8000
  description: 本機開發伺服器